Add lint rule to prefer stateless function components
After some internal discussion, we decided to use function components
(with hooks) unless you have a very good reason to use a class
component.

This adds [`react/prefer-stateless-function` ESLint rule][1] which
gives a lint error if you try to use a class component that should be
a stateless function component.

This doesn't go all the way to achieving our stated goal—it won't
error if you use class components that have state—but I think this is
a useful way to enforce a piece of our style rule.

[1]: https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/prefer-stateless-function.md