Update <Link> styles to be more generic
This updates the `<Link>` component's styles.

Before this change, they were only used in the header, so they had
styles that were specific to the header. We'll want to use this
component in more places, so I removed the header-specific styles and
moved them into the `<Header>` component.

This was useful while working on the project page (see [#14][1]).

[1]: https://github.com/climateconnect/climateconnect/issues/14