Fix frontend environment variables for Google Cloud
Google Cloud has us define environment variables in `app.yaml`. However,
these environment variables are not available at build time, only at
runtime.

To get around this, we read the variables from `app.yaml` during the
build.

This is a bit of a hack but I hope it's straightforward. If we change
our production build setup, we should revisit this solution.